Legends: (!) Important, take note. (-) Fixes (+) New Features/Items
<b>(!) I have edited create table statement in RationaleDBCreate and changed "GENERATED ALWAYS AS IDENTITY" to "GENERATED BY DEFAULT AS IDENTITY" so that it is possible
to manually specify the id of the entities.</b>

<b>(!)We're now creating the database via the XML file if such file exist. (ratDBCreate.xml)</b>


(+)Added ratDBCreate.xml file.

(+)Fixed the export XML problem in which it has potential to add duplicated entry of the same pattern decision.

(-)Fixed the import XML problem when Apache Derby database is in use.

(-)We're now closing the statements for XML import so that I don't need to "flush table" after it's done.

(+)If there is a duplicated entry for some reason, we will ignore it instead of printStackTrace.