Coding of generate tactic.
Minimal test done on generate tactic.

Fixed sync problem in case editor was opened and a tactic association was deleted.

Fixed the tree not updating when tactic has changed its positive attribute.

Added capability of adding a tactic by right clicking on a category. The tactic added will automatically be assigned to the tactic category.