(!) Important, take note, (+) Added features (-) Bug fixes (PL) Pattern Library (TL) Tactic Library (RE) Rationale Explorer (RT) Rationale Task List
(!) Alternative.java now has generatePattern method always using contributing matching algorithm. The code supporting exact and not matching are deleted.
(!) The signature for generatePattern now only takes the paramter (Decision dec). It no longer supports (int matchingMethod)
(!) AlternativePatternInference and SelectCandidatePattern classes are now deprecated
(!) The scoring of the pattern generation wizard is done in a tree instead of using double digit number. The user may use their own judgement to determine which is more important.

(-) (PL) Fixed the "amount" in pattern library to be 5 for all arguments.
(-) (PL) The plausibility is always HIGH, unless it is a potential match. If it is potential match, the palusbility is low
(-) (TL/RE) The decisions may be generated outside the subtree of the decision of the pattern.
(+) (RE) Added support for "Complicates" and "Faciliates" for arguments of type Alternative.
(-) (TL/RE) Changed the arguments generated for specific patterns to use their alternative.
(-) (PL/RE) The pattern generation wizard no longer have a page on Contribution or Exact matching.